Time: 0 ms (100.00%) | Memory: 6.7 MB (8.56%) - LeetSync

+++ 342-power-of-four/power-of-four.cpp
@@ -5,7 +5,7 @@ public:
             return false;
         if (n == 1)
             return true;
-       double x=log (n)/log (4);
+       float x=log (n)/log (4);
        return x==(int)x;
